requirements.txt
Django==2.2.4
django-paypal==1.0.0
